feat(localstore): add migration for batch-index
