+++ pkg/localstore/schema.go
@@ -16,16 +16,9 @@
 
 package localstore
 
-// The DB schema we want to use. The actual/current DB schema might differ
-// until migrations are run.
-var DbSchemaCurrent = DbSchemaYuj
+// DBSchemaCode is the first bee schema identifier.
+const DBSchemaCode = "code"
 
-// There was a time when we had no schema at all.
-const DbSchemaNone = ""
-
-// DbSchemaCode is the first bee schema identifier
-const DbSchemaCode = "code"
-
-// DbSchemaYuj is the bee schema indentifier for storage incentives
-// initial iteration.
-const DbSchemaYuj = "yuj"
+// DBSchemaCurrent represents the DB schema we want to use.
+// The actual/current DB schema might differ until migrations are run.
+var DBSchemaCurrent = DBSchemaBatchIndex
